+++ src/bl64-arc.bash
@@ -24,8 +24,7 @@ function bl64_arc_run_unzip() {
     bl64_check_parameters_none "$#" &&
     bl64_check_command "$BL64_ARC_CMD_UNZIP" || return $?
 
-  bl64_msg_lib_verbose_enabled && verbosity='-q'
-  bl64_dbg_lib_command_enabled && verbosity=' '
+  bl64_msg_lib_verbose_enabled && verbosity=' '
 
   bl64_arc_blank_unzip
 
@@ -80,13 +79,13 @@ function bl64_arc_run_tar() {
     bl64_check_command "$BL64_ARC_CMD_TAR" ||
     return $?
 
-  bl64_dbg_lib_command_enabled && debug="$BL64_ARC_SET_TAR_VERBOSE"
+  bl64_msg_lib_verbose_enabled && debug="$BL64_ARC_SET_TAR_VERBOSE"
 
   bl64_dbg_lib_trace_start
   # shellcheck disable=SC2086
   "$BL64_ARC_CMD_TAR" \
-    $debug \
-    "$@"
+    "$@" \
+    $debug
   bl64_dbg_lib_trace_stop
 }